requirements.txt
Flask==2.3.3
tensorflow==2.13.0
numpy==1.24.4
Werkzeug==2.3.3
Pillow